feat(hero): setup Lottie animation in HeroVisual component
- Create new HeroVisual component to isolate animation logic
- Move Lottie animation rendering and interactivity from Hero.js
- Add scroll-responsive playback and mouse interaction tracking
- Import animation from ./public/lottiefiles/ea6225fb-859a-4eb1-9706-1f52f789d436.json
- Simplify Hero.js to focus on content layout
- Maintain theme colour application and LiquidGL integration
